Add files via upload
It is my first result for my personal web page project. I tried to create an elegant and simple design. I used only HTML5 and CSS for this result. 

What will I add to the second result? 
- Animations and SCSS
- Bootstrap

Which sources using for this exercise?
- https://www.udemy.com/course/komple-web-developer-kursu/learn/lecture/17801896?start=15#overview
- https://www.youtube.com/watch?v=HBz6z9PDemE&list=PL3AsBN9JULajo7Vfq1haOdgwkGkT0UkuF

For the first result, I used these sources. When this project going to improve and to be complicated I'm going to use extra sources of course. I'll add all sources below ("add an optional extended description" section) when I use them in my project.